third_party/blink/web_tests/http/tests/cookies/js-set-null-expected.txt

Setting document.cookie to null should set a cookie named 'null'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.cookie is 'null'
PASS cookie is 'null'.
PASS successfullyParsed is true

TEST COMPLETE
